Update data.yml
I edited bios to try and make them a little more timeless. I'd love to see some links within your content, especially where you're adding personality. All of the bios would really benefit from something that says, "The apprenticeship was so worth it, because..." Something along that idea. Since this page is supposed to say who you are, but also meet a main goal of educating and promoting apprenticeships—it would be good to have you say it was a great experience.