Update 01-home.md
I love and also question this "so that one day the apprentice can become the master." I think it could be just fine, but might be something to ponder/see if you have any other ideas. Think about what you're gaining from the apprenticeship, and see if that would lead you toward something else. If not, then I think it works.